+++ bootstrap.php
@@ -57,15 +57,22 @@ function cmb2_bootstrap() {
 	 */
 	foreach ( CMB2_Boxes::get_all() as $cmb ) {
 
-		if ( $cmb->prop( 'hookup' ) ) {
-			$hookup = new CMB2_hookup( $cmb );
-			$hookup->universal_hooks();
-		}
-
-		if ( $cmb->prop( 'show_in_rest' ) && function_exists( 'rest_get_server' ) ) {
-			$rest = new CMB2_REST( $cmb );
-			$rest->universal_hooks();
-		}
+		/**
+		 * Initiates the box "hookup" into WordPress.
+		 *
+		 * Unless the 'hookup' box property is `false`, the box will be hooked in as
+		 * a post/user/comment/option/term box.
+		 *
+		 * And if the 'show_in_rest' box property is set, the box will be hooked
+		 * into the CMB2 REST API.
+		 *
+		 * The dynamic portion of the hook name, $cmb->cmb_id, is the box id.
+		 *
+		 * @since 2.2.6
+		 *
+		 * @param array $cmb The CMB2 object to hookup.
+		 */
+		do_action( "cmb2_init_hookup_{$cmb->cmb_id}", $cmb );
 	}
 
 	/**